bookstore/requirements.txt
datetime
jieba
simplejson
psycopg2
sqlalchemy
lxml
codecov
coverage
flask
pre-commit
pytest
PyJWT
requests

